requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
torch==1.9.0
torchvision==0.10.0
basicsr==1.3.4.0
realesrgan==0.2.5.0
Pillow==8.3.1